Add updated database spec to Dart frontend and add point_finder utility on backend to propagate changes to positions and calculate points
