+++ functions/src/authentication.ts
@@ -1,0 +1,106 @@
+import * as functions from "firebase-functions";
+import {hashAdminToken, randomCode} from "./helpers";
+import * as admin from "firebase-admin";
+import {v4 as uuid} from "uuid";
+import {createTransport} from "nodemailer";
+
+interface AdminSignInData {
+    token?: string;
+}
+export const getAdminSignInToken = functions
+    .region('europe-west2')
+    .https.onCall(async (data: AdminSignInData) => {
+        if (!data.token) {
+            return null;
+        }
+
+        const hashedToken = hashAdminToken(data.token);
+        const response = await admin.firestore().collection('tokens')
+            .doc(hashedToken)
+            .get();
+
+        if (!response.exists) {
+            return null;
+        }
+
+        return await admin.auth().createCustomToken(hashedToken, {
+            adminEnabled: true,
+        });
+    });
+
+interface SendEmailData {
+    email?: string;
+}
+export const sendEmail = functions
+    .region('europe-west2')
+    .https.onCall(async (data: SendEmailData): Promise<string | undefined> => {
+        if (!data.email || !data.email.endsWith('@mgs.org')) {
+            return;
+        }
+
+        const code = randomCode(6);
+        const sessionId = uuid();
+        await admin.firestore()
+            .collection('setup_codes')
+            .doc(sessionId)
+            .set({
+                code,
+                createdAt: admin.firestore.Timestamp.now(),
+            });
+
+        const transporter = createTransport({
+            host: functions.config().smtp.host,
+            port: parseInt(functions.config().smtp.port),
+            auth: {
+                user: functions.config().smtp.user,
+                pass: functions.config().smtp.pass,
+            }
+        });
+
+        try {
+            await transporter.sendMail({
+                from: functions.config().smtp.sender,
+                sender: functions.config().smtp.sender,
+                to: data.email,
+                replyTo: functions.config().smtp.reply_to,
+                subject: 'Confirm your email for MyMGS',
+                text: `Your MyMGS setup code is ${code}`,
+            });
+        } catch (e) {
+            return;
+        }
+
+        return sessionId;
+    });
+
+interface ConfirmEmailData {
+    sessionId?: string;
+    code?: string;
+}
+export const confirmEmail = functions
+    .region('europe-west2')
+    .https.onCall(async (data: ConfirmEmailData): Promise<string | undefined> => {
+        if (!data.sessionId || !data.code) {
+            return;
+        }
+
+        const sessionResponse = await admin.firestore()
+            .collection('setup_codes')
+            .doc(data.sessionId)
+            .get();
+
+        if (!sessionResponse.exists) {
+            return;
+        }
+
+        if (sessionResponse.data()?.code !== data.code) {
+            return;
+        }
+
+        // delete the setup session for security
+        await sessionResponse.ref.delete();
+
+        // generate a new uuid to avoid any accidental path of association with the user's email address
+        const userId = uuid();
+        return admin.auth().createCustomToken(userId);
+    });
