+++ functions/src/index.ts
@@ -1,166 +1,10 @@
-import * as functions from 'firebase-functions';
 import * as admin from 'firebase-admin';
-import { v4 as uuid } from 'uuid';
-import {hashAdminToken, randomCode} from './helpers';
-import { createTransport } from 'nodemailer';
+import { getAdminSignInToken, confirmEmail, sendEmail } from './authentication';
+import { sendNotification } from './notifications';
+import { sdPropagateScore } from './sportsday/sportsday';
 
 admin.initializeApp();
 
-interface NotificationData {
-    topic: string;
-    title: string;
-    caption: string;
-    imageUrl: string;
-    resourceType: number;
-    resourceId: string;
-    token: string;
-}
-
-export const sendNotification = functions
-    .region('europe-west2')
-    .https.onCall(async (data: Partial<NotificationData>): Promise<boolean> => {
-        if (!data.token || !data.title || !data.caption || !data.topic) {
-            return false;
-        }
-
-        let verifiedToken: admin.auth.DecodedIdToken;
-        try {
-            verifiedToken = await admin.auth().verifyIdToken(data.token);
-        } catch (e) {
-            return false;
-        }
-
-        const userResponse = await admin.firestore()
-            .collection('tokens')
-            .doc(verifiedToken.uid)
-            .get();
-
-        const user = userResponse.data();
-        if (!userResponse.exists || !user) return false;
-        if (!user.perms.includes('notifications')) return false;
-
-        const payload = {
-            topic: data.topic,
-            title: data.title,
-            body: data.caption,
-        } as {
-            [key: string]: any;
-        };
-
-        if (data.imageUrl) {
-            payload.imageUrl = data.imageUrl;
-        }
-
-        if (data.resourceId && data.resourceType) {
-            payload.resourceType = data.resourceType.toString();
-            payload.resourceId = data.resourceId;
-        }
-
-        await admin.messaging().send({
-            topic: data.topic,
-            data: payload,
-        });
-        return true;
-    });
-
-interface AdminSignInData {
-    token?: string;
-}
-
-export const getAdminSignInToken = functions
-    .region('europe-west2')
-    .https.onCall(async (data: AdminSignInData) => {
-        if (!data.token) {
-            return null;
-        }
-
-        const hashedToken = hashAdminToken(data.token);
-        const response = await admin.firestore().collection('tokens')
-            .doc(hashedToken)
-            .get();
-
-        if (!response.exists) {
-            return null;
-        }
-
-        return await admin.auth().createCustomToken(hashedToken, {
-            adminEnabled: true,
-        });
-    });
-
-interface SendEmailData {
-    email?: string;
-}
-export const sendEmail = functions
-    .region('europe-west2')
-    .https.onCall(async (data: SendEmailData): Promise<string | undefined> => {
-        if (!data.email || !data.email.endsWith('@mgs.org')) {
-            return;
-        }
-
-        const code = randomCode(6);
-        const sessionId = uuid();
-        await admin.firestore()
-            .collection('setup_codes')
-            .doc(sessionId)
-            .set({
-                code,
-                createdAt: admin.firestore.Timestamp.now(),
-            });
-
-        const transporter = createTransport({
-            host: functions.config().smtp.host,
-            port: parseInt(functions.config().smtp.port),
-            auth: {
-                user: functions.config().smtp.user,
-                pass: functions.config().smtp.pass,
-            }
-        });
-
-        try {
-            await transporter.sendMail({
-                from: functions.config().smtp.sender,
-                sender: functions.config().smtp.sender,
-                to: data.email,
-                replyTo: functions.config().smtp.reply_to,
-                subject: 'Confirm your email for MyMGS',
-                text: `Your MyMGS setup code is ${code}`,
-            });
-        } catch (e) {
-            return;
-        }
-
-        return sessionId;
-    });
-
-interface ConfirmEmailData {
-    sessionId?: string;
-    code?: string;
-}
-export const confirmEmail = functions
-    .region('europe-west2')
-    .https.onCall(async (data: ConfirmEmailData): Promise<string | undefined> => {
-        if (!data.sessionId || !data.code) {
-            return;
-        }
-
-        const sessionResponse = await admin.firestore()
-            .collection('setup_codes')
-            .doc(data.sessionId)
-            .get();
-
-        if (!sessionResponse.exists) {
-            return;
-        }
-
-        if (sessionResponse.data()?.code !== data.code) {
-            return;
-        }
-
-        // delete the setup session for security
-        await sessionResponse.ref.delete();
-
-        // generate a new uuid to avoid any accidental path of association with the user's email address
-        const userId = uuid();
-        return admin.auth().createCustomToken(userId);
-    });
+export {sendNotification};
+export {getAdminSignInToken, confirmEmail, sendEmail};
+export {sdPropagateScore};
