+++ lib/data_classes/sportsday/event.dart
@@ -5,13 +5,15 @@ part 'event.g.dart';
 @JsonSerializable()
 class Event {
   String id;
-  String name;
-  int subEventCount;
+  int yearGroup;
+  int subEvent;
+  String eventGroupId;
 
   Event({
     required this.id,
-    required this.name,
-    required this.subEventCount,
+    required this.subEvent,
+    required this.eventGroupId,
+    required this.yearGroup,
   });
   factory Event.fromJson(Map<String, dynamic> json) => _$EventFromJson(json);
   Map<String, dynamic> toJson() => _$EventToJson(this);
